test: add buildComparisonTrees to test utils
